+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Program argument that is used as the seed of the random number generator
+
+### Fixed
+- Restaurant example which still had tests within it
 
 ## [1.2.1] - 2018-10-22
 ### Changed